rename issuer to cert-issuer and remove bad links
