+++ conf_regtest.ini
@@ -3,7 +3,7 @@ revocation_address = <revocation-address>
 
 wallet_connector_type = bitcoind
 broadcaster_type = bitcoind
-usb_name=/etc/issuer/
+usb_name=/etc/cert-issuer/
 key_file = pk_issuer.txt
 
 skip_wifi_check
