+++ docs/Makefile
@@ -96,9 +96,9 @@ qthelp:
 	@echo
 	@echo "Build finished; now you can run "qcollectiongenerator" with the" \
 	      ".qhcp project file in $(BUILDDIR)/qthelp, like this:"
-	@echo "# qcollectiongenerator $(BUILDDIR)/qthelp/issuer.qhcp"
+	@echo "# qcollectiongenerator $(BUILDDIR)/qthelp/cert-issuer.qhcp"
 	@echo "To view the help file:"
-	@echo "# assistant -collectionFile $(BUILDDIR)/qthelp/issuer.qhc"
+	@echo "# assistant -collectionFile $(BUILDDIR)/qthelp/cert-issuer.qhc"
 
 .PHONY: applehelp
 applehelp:
@@ -115,8 +115,8 @@ devhelp:
 	@echo
 	@echo "Build finished."
 	@echo "To view the help file:"
-	@echo "# mkdir -p $$HOME/.local/share/devhelp/issuer"
-	@echo "# ln -s $(BUILDDIR)/devhelp $$HOME/.local/share/devhelp/issuer"
+	@echo "# mkdir -p $$HOME/.local/share/devhelp/cert-issuer"
+	@echo "# ln -s $(BUILDDIR)/devhelp $$HOME/.local/share/devhelp/cert-issuer"
 	@echo "# devhelp"
 
 .PHONY: epub
