+++ docs/docker.md
@@ -7,14 +7,14 @@ experimenting only.
 1. Clone the repo:
 
     ```
-    git clone https://github.com/digital-certificates/issuer.git
+    git clone https://github.com/digital-certificates/cert-issuer.git
     ```
 
-2. From a command line in issuer dir, build your docker container
+2. From a command line in cert-issuer dir, build your docker container
 
     ```
-    cd issuer
-    docker build -t ml/issuer:1.0 .
+    cd cert-issuer
+    docker build -t ml/cert-issuer:1.0 .
     ```
 
 3. Read before running!
@@ -25,13 +25,13 @@ experimenting only.
 
     ```
     docker ps -l
-    docker commit <container for your ml/issuer> my_cert_issuer
+    docker commit <container for your ml/cert-issuer> my_cert_issuer
     ```
 
 4. When you're ready to run:
 
     ```
-    docker run -it ml/issuer:1.0 bash
+    docker run -it ml/cert-issuer:1.0 bash
 
     ```
 
@@ -52,15 +52,15 @@ Ensure your docker image is running and bitcoind process is started
 
     ```
     issuer=`bitcoin-cli getnewaddress`
-    sed -i.bak "s/<issuing-address>/$issuer/g" /etc/issuer/conf.ini
-    bitcoin-cli dumpprivkey $issuer > /etc/issuer/pk_issuer.txt
+    sed -i.bak "s/<issuing-address>/$issuer/g" /etc/cert-issuer/conf.ini
+    bitcoin-cli dumpprivkey $issuer > /etc/cert-issuer/pk_issuer.txt
     ```
 2. Create a 'revocation address' and save the output as follows. Note that we don't need to save this
 corresponding private key for testing issuing certificates
 
     ```
     revocation=`bitcoin-cli getnewaddress`
-    sed -i.bak "s/<revocation-address>/$revocation/g" /etc/issuer/conf.ini
+    sed -i.bak "s/<revocation-address>/$revocation/g" /etc/cert-issuer/conf.ini
     ```
 
 3. Don't forget to save snapshots so you don't lose your work (see step 3 of client setup)
@@ -81,16 +81,16 @@ corresponding private key for testing issuing certificates
     b. Send the money to your issuing address -- note bitcoin-cli's standard denomination is bitcoins not satoshis! In our
     app, the standard unit is satoshis.
     ```
-    bitcoin-cli sendtoaddress moH7X29kt5T8fbxTCjoxYzjfLeMR56Ju94 5  << bitcoins not satoshi!!!!
+    bitcoin-cli sendtoaddress $issuer 5  << bitcoins not satoshi!!!!
     ```
 
 3. Run
-TODO: there is an issuer in the Dockerfile and/or setup.py. The pip install here shouldn't be required.
+TODO: there is an cert-issuer in the Dockerfile and/or setup.py. The pip install here shouldn't be required.
 For now, these get it running:
 
     ```
-    source /issuer/env/bin/activate
-    cd issuer
+    source /cert-issuer/env/bin/activate
+    cd cert-issuer
     pip install .
-    python issuer -c /etc/issuer/conf.ini
+    cert-issuer -c /etc/cert-issuer/conf.ini
     ```
