+++ docs/modules.rst
@@ -1,7 +1,7 @@
-issuer
+cert_issuer
 ==================
 
 .. toctree::
    :maxdepth: 4
 
-   issuer
+   cert_issuer
